move to single agent struct, remove agent trait
