+++ crates/goose/src/agents/mod.rs
@@ -1,10 +1,8 @@
 mod agent;
 pub mod capabilities;
 pub mod extension;
-mod goose_agent;
 mod types;
 
 pub use agent::{Agent, SessionConfig};
 pub use capabilities::Capabilities;
 pub use extension::ExtensionConfig;
-pub use goose_agent::GooseAgent;
